refactor: organize futures into seperate file
