+++ src/lib.rs
@@ -1,6 +1,7 @@
 #![forbid(unsafe_code)]
 
 pub(crate) mod app;
+pub(crate) mod futures;
 pub(crate) mod screen;
 pub(crate) mod utils;
 pub use app::App;
